Revert "fix(posts-form): Fixes default number field"

+++ app/main/posts/modify/post-data-editor.directive.js
@@ -259,8 +259,7 @@ function PostDataEditorController(
                             $scope.post.values[attr.key] = [null];
                         }
                     }  else if (attr.input === 'number') {
-                        // parseFloat to handle decimals too.
-                        $scope.post.values[attr.key] = [parseFloat(attr.default)];
+                        $scope.post.values[attr.key] = [parseInt(attr.default)];
                     } else if (attr.input === 'date' || attr.input === 'datetime') {
                         $scope.post.values[attr.key] = attr.default ? [new Date(attr.default)] : [new Date()];
                     } else {
